Renamed dokomoforms/handlers/view to dokomoforms/handlers/user [#107206058]
